requirements.txt
django
django-tinymce
fontawesomefree
django-crispy-forms
crispy-bootstrap4
django-recaptcha
pillow
six
python-dotenv
